Reformat/restyle files according to rules in .pre-commit-config.yaml
Reformatted the files in the repository according to rules in a new
file ".pre-commit-config.yaml".  Have to install pre-commit either
through pip or conda and then "pre-commit install" to install the
git hooks.

The pre-commit lints or reformats many file types by using
separate linters and autoformaters that are installed as part of
pre-commit.

Uses "ruff" for Python files, but presently configured only for
autoformatting.  Turned off "ruff" linting because it found over
300 problems, some of which there weren't easy answers for.